Move lots of logic out of the Schedule file into separate files
